+++ type-generator/README.md
@@ -1,0 +1,8 @@
+# Type Generator
+
+The Marathon type generator, as the name suggests, generates types base on [RAML](https://raml.org/) API specifications.
+
+## Package structure
+
+* `marathon.raml.ir` holds the intermediate representation.
+* `marathon.raml.backend` hold the code generation.
