+++ type-generator/src/main/scala/mesosphere/raml/RamlTypeGenerator.scala
@@ -2,6 +2,8 @@ package mesosphere.raml
 
 import treehugger.forest._
 import definitions._
+import mesosphere.raml.backend._
+import mesosphere.raml.ir.{Constraint, GeneratedClass, EnumT, FieldT, ObjectT, StringT, UnionT}
 import org.raml.v2.api.RamlModelResult
 import org.raml.v2.api.model.v10.api.Library
 import org.raml.v2.api.model.v10.datamodel._
@@ -12,81 +14,6 @@ import scala.collection.JavaConverters._
 import scala.collection.immutable.Seq
 
 object RamlTypeGenerator {
-  val AdditionalProperties = "additionalProperties"
-  val baseTypeTable: Map[String, Symbol] =
-    Map(
-      "string" -> StringClass,
-      "int8" -> ByteClass,
-      "int16" -> ShortClass,
-      "int32" -> IntClass,
-      "integer" -> IntClass,
-      "int64" -> LongClass,
-      "long" -> LongClass,
-      "float" -> FloatClass,
-      "double" -> DoubleClass,
-      "boolean" -> BooleanClass,
-      "date-only" -> RootClass.newClass("java.time.LocalDate"),
-      "time-only" -> RootClass.newClass("java.time.LocalTime"),
-      "datetime-only" -> RootClass.newClass("java.time.LocalDateTime"),
-      "datetime" -> RootClass.newClass("java.time.OffsetDateTime"),
-      "RamlGenerated" -> RootClass.newClass("RamlGenerated"),
-      "RamlConstraints" -> RootClass.newClass("RamlConstraints")
-    )
-
-  val builtInTypes = Set(
-    "Byte",
-    "Short",
-    "Int",
-    "Long",
-    "Float",
-    "Double",
-    "Boolean",
-    "String",
-    "java.time.LocalDate",
-    "java.time.LocalTime",
-    "java.time.LocalDateTime",
-    "java.time.OffsetDateTime"
-  )
-
-  val TryClass = RootClass.newClass("scala.util.Try")
-
-  val SeqClass = RootClass.newClass("scala.collection.immutable.Seq")
-  val SetClass = RootClass.newClass("Set")
-  val IterableClass = RootClass.newClass("Iterable")
-
-  def TYPE_SEQ(typ: Type): Type = SeqClass TYPE_OF typ
-
-  val PlayJsonFormat = RootClass.newClass("play.api.libs.json.Format")
-
-  def PLAY_JSON_FORMAT(typ: Type): Type = PlayJsonFormat TYPE_OF typ
-
-  val PlayJsonResult = RootClass.newClass("play.api.libs.json.JsResult")
-
-  def PLAY_JSON_RESULT(typ: Type): Type = PlayJsonResult TYPE_OF typ
-
-  val PlayJson = RootClass.newClass("play.api.libs.json.Json")
-  val PlayJsValue = RootClass.newClass("play.api.libs.json.JsValue")
-  val PlayJsString = RootClass.newClass("play.api.libs.json.JsString")
-  val PlayJsObject = RootClass.newClass("play.api.libs.json.JsObject")
-  val PlayJsArray = RootClass.newClass("play.api.libs.json.JsArray")
-  val PlayValidationError = RootClass.newClass("play.api.libs.json.JsonValidationError")
-  val PlayJsError = RootClass.newClass("play.api.libs.json.JsError")
-  val PlayJsSuccess = RootClass.newClass("play.api.libs.json.JsSuccess")
-  val PlayReads = RootClass.newClass("play.api.libs.json.Reads")
-  def PLAY_JSON_READS(typ: Type): Type = PlayReads TYPE_OF typ
-  val PlayWrites = RootClass.newClass("play.api.libs.json.Writes")
-  def PLAY_JSON_WRITES(typ: Type): Type = PlayWrites TYPE_OF typ
-  val PlayPath = RootClass.newClass("play.api.libs.json.JsPath")
-
-  val PlayJsNull = REF("play.api.libs.json.JsNull")
-
-  val NoScalaFormat = "format: OFF"
-
-  def camelify(name: String): String = name.toLowerCase.capitalize
-
-  def underscoreToCamel(name: String) = "(/|_|\\,)([a-z\\d])".r.replaceAllIn(name, { m =>
-    m.group(2).toUpperCase()
-  })
 
   def enumName(s: StringTypeDeclaration, default: Option[String] = None): String = {
     s.annotations().asScala.find(_.name() == "(pragma.scalaType)").fold(default.getOrElse(s.name()).capitalize) { annotation =>
@@ -104,11 +31,6 @@ object RamlTypeGenerator {
     } else {
       o.name() -> Some(o.`type`())
     }
-  }
-
-  def scalaFieldName(name: String): String = {
-    if (name.contains("-")) s"`$name`"
-    else name
   }
 
   def isUpdateType(o: ObjectTypeDeclaration): Boolean =
@@ -162,480 +84,6 @@ object RamlTypeGenerator {
       }
     }
     build(types, baseTypeTable)
-  }
-
-  sealed trait GeneratedClass {
-    val name: String
-
-    def toTree(): Seq[Tree]
-  }
-
-  case class EnumT(name: String, values: Set[String], default: Option[String], comments: Seq[String]) extends GeneratedClass {
-    val sortedValues = values.toVector.sorted
-    override def toString: String = s"Enum($name, $values)"
-
-    override def toTree(): Seq[Tree] = {
-      val baseTrait = TRAITDEF(name) withParents("Product", "Serializable", "RamlGenerated") withFlags Flags.SEALED := BLOCK(
-        VAL("value", StringClass),
-        DEF("toString", StringClass) withFlags Flags.OVERRIDE := REF("value")
-      )
-
-      val enumObjects = sortedValues.map { enumValue =>
-        CASEOBJECTDEF(underscoreToCamel(camelify(enumValue))) withParents name := BLOCK(
-          VAL("value") := LIT(enumValue)
-        )
-      }
-
-      val patternMatches = sortedValues.map { enumValue =>
-        CASE(LIT(enumValue.toLowerCase)) ==> REF(underscoreToCamel(camelify(enumValue)))
-      }
-
-      val playWildcard = CASE(WILDCARD) ==>
-        (REF(PlayJsError) APPLY (REF(PlayValidationError) APPLY(LIT("error.unknown.enum.literal"), LIT(s"$name (${sortedValues.mkString(", ")})"))))
-      val playPatternMatches = sortedValues.map { enumValue =>
-        CASE(LIT(enumValue.toLowerCase)) ==> (REF(PlayJsSuccess) APPLY REF(underscoreToCamel(camelify(enumValue))))
-      }
-
-      val playJsonFormat = (OBJECTDEF("playJsonFormat") withParents PLAY_JSON_FORMAT(name) withFlags Flags.IMPLICIT) := BLOCK(
-        DEF("reads", PLAY_JSON_RESULT(name)) withParams PARAM("json", PlayJsValue) := {
-          REF("json") MATCH(
-            CASE(REF(PlayJsString) UNAPPLY ID("s")) ==> (REF("s") DOT "toLowerCase" MATCH (playPatternMatches ++ Vector(playWildcard))),
-            playWildcard)
-        },
-        DEF("writes", PlayJsValue) withParams PARAM("o", name) := {
-          REF(PlayJsString) APPLY (REF("o") DOT "value")
-        }
-      )
-
-      val obj = OBJECTDEF(name) := BLOCK(
-        enumObjects ++ Seq(
-          playJsonFormat,
-          VAL("StringToValue") withType(TYPE_MAP(StringClass, name)) withFlags(Flags.PRIVATE) := REF("Map") APPLY(sortedValues.map { enumValue =>
-            TUPLE(LIT(enumValue), REF(underscoreToCamel(camelify(enumValue))))
-          }),
-          DEF("all", IterableClass TYPE_OF name) := REF("StringToValue") DOT "values",
-          DEF("fromString", TYPE_OPTION(name)) withParams(PARAM("v", StringClass)) := REF("StringToValue") DOT "get" APPLY(REF("v"))
-        ) ++ default.map { defaultValue =>
-          VAL("DefaultValue") withType(name) := REF(underscoreToCamel(camelify(defaultValue)))
-        }
-      )
-      Seq(baseTrait.withDoc(comments), obj)
-    }
-  }
-
-  sealed trait Constraint[C] { self =>
-    val name: String
-    val constraint: C
-    val constraintToValue: C => Tree = { (c: C) => LIT(c) } // decent assumption for built-ins, probably not much else
-
-    /** false indicates custom, non-play constraint implementations that we've implemented as part of this generator */
-    val builtIn: Boolean
-
-    /** @return a code gen expression that represents a playJS reads validation */
-    def validate(): Tree
-
-    /** a code gen expression for a field that represents the constraint limit */
-    val limitField: Option[Tree] = None
-
-    /** decorate this constraint with a `limitField` implementation */
-    def copyWith(lf: Option[Tree] = None): Constraint[C]
-
-    def withFieldLimit(f: FieldT): Constraint[C] = {
-      val fieldName = scalaFieldName(underscoreToCamel(camelify(s"constraint_${f.rawName}_${name}".replace("-", "_"))))
-      val limit = (VAL(fieldName) := constraintToValue(constraint))
-      copyWith(Option(limit))
-    }
-  }
-
-  object Constraint {
-
-    // built-in playJS validators
-
-    def MaxLength(len: Integer) = Constraint("maxLength", len, builtIn = true) { REF(_) APPLYTYPE StringClass APPLY(_) }
-    def MinLength(len: Integer) = Constraint("minLength", len, builtIn = true) { REF(_) APPLYTYPE StringClass APPLY(_) }
-
-    def Pattern(p: String) = Constraint("pattern", p, builtIn = true, (c: String) => LIT(c) DOT "r") { REF(_) APPLY(_) }
-
-    def MaxItems(len: Integer, t: Type) = Constraint("maxLength", len, builtIn = true) { REF(_) APPLYTYPE t APPLY(_) }
-    def MinItems(len: Integer, t: Type) = Constraint("minLength", len, builtIn = true) { REF(_) APPLYTYPE t APPLY(_) }
-
-    def Max(v: Number, t: Type) = Constraint("max", v, builtIn = true) { REF(_) APPLYTYPE t APPLY(_) }
-    def Min(v: Number, t: Type) = Constraint("min", v, builtIn = true) { REF(_) APPLYTYPE t APPLY(_) }
-
-    // custom validator implementations follow
-
-    def KeyPattern(p: String, mapValType: Type) =
-      Constraint("keyPattern", p, builtIn = false, (c: String) => LIT(c) DOT "r") { REF(_) APPLYTYPE mapValType APPLY(_) }
-
-    case class BasicConstraint[C](
-      override val name: String,
-      override val constraint: C,
-      override val constraintToValue: C => Tree,
-      val validateFunc: (String, Tree) => Tree,
-      override val builtIn: Boolean,
-      override val limitField: Option[Tree] = None) extends Constraint[C] {
-      override def validate(): Tree = validateFunc(name, constraintToValue(constraint))
-      override def copyWith(lf: Option[Tree] = limitField): Constraint[C] = copy(limitField = lf)
-    }
-
-    def apply[C](n: String, c: C, builtIn: Boolean, c2v: C => Tree = { (c: C) => LIT(c) })(f: (String, Tree) => Tree): Constraint[C] =
-      new BasicConstraint(n, c, c2v, f, builtIn)
-
-    implicit class Constraints(val c: Seq[Constraint[_]]) extends AnyVal {
-      def validate(exp: Tree): Tree = {
-        if (c.isEmpty) {
-          exp
-        } else {
-          @tailrec
-          def buildChain(constraints: List[Constraint[_]], chain: Tree): Tree = constraints match {
-            case Nil => chain
-            case c :: rs => buildChain(rs, chain INFIX("keepAnd", c.validate()))
-          }
-
-          exp APPLY buildChain(c.tail.to[List], c.head.validate())
-        }
-      }
-
-    }
-
-    implicit class AllConstraints(val c: Seq[Seq[Constraint[_]]]) extends AnyVal {
-      def requiredImports: Seq[Tree] = {
-        val flattened = c.flatten
-        if (flattened.isEmpty) {
-          Nil
-        } else {
-          Seq(
-            Option(IMPORT(PlayReads DOT "_")),
-            if (c.exists(_.size > 1)) Option(IMPORT("play.api.libs.functional.syntax._")) else None,
-            flattened.find(!_.builtIn).map(_ => IMPORT("RamlConstraints._"))
-          )
-        }.flatten
-      }
-    }
-  }
-
-  case class FieldT(rawName: String, `type`: Type, comments: Seq[String], constraints: Seq[Constraint[_]], required: Boolean,
-                    default: Option[String], repeated: Boolean = false, forceOptional: Boolean = false, omitEmpty: Boolean = false) {
-
-    val name = scalaFieldName(rawName)
-    override def toString: String = s"$name: ${`type`}"
-
-    lazy val paramTypeValue: Option[(Type, Tree)] = {
-      if ((required || default.isDefined) && !forceOptional) {
-        defaultValue.map { d => `type` -> d }
-      } else {
-        Option(
-          if (repeated && !forceOptional) {
-            val typeName = `type`.toString()
-            if (typeName.startsWith("Map")) {
-              `type` -> (REF("Map") DOT "empty")
-            } else {
-              if (typeName.startsWith("Set")) {
-                `type` -> (REF("Set") DOT "empty")
-              } else {
-                `type` -> NIL
-              }
-            }
-          } else {
-            TYPE_OPTION(`type`) -> NONE
-          }
-        )
-      }
-    }
-
-    lazy val param: treehugger.forest.ValDef =
-      paramTypeValue.fold { PARAM(name, `type`).tree } { case (pType, pValue) => PARAM(name, pType) := pValue }
-
-    lazy val comment: String = if (comments.nonEmpty) {
-      val lines = comments.flatMap(_.lines)
-      s"@param $name ${lines.head} ${if (lines.tail.nonEmpty) "\n  " else ""}${lines.tail.mkString("\n  ")}"
-    } else {
-      ""
-    }
-
-    val defaultValue: Option[Tree] = default.map { d =>
-      `type`.toString() match {
-        case "Byte" => LIT(d.toByte)
-        case "Short" => LIT(d.toShort)
-        case "Int" => LIT(d.toInt)
-        case "Long" => LIT(d.toLong)
-        case "Float" => LIT(d.toFloat)
-        case "Double" => LIT(d.toDouble)
-        case "Boolean" => LIT(d.toBoolean)
-        case "String" => LIT(d)
-        // hopefully this is actually an enum
-        case _ => (`type` DOT underscoreToCamel(camelify(d))).tree
-      }
-    }
-
-    val playReader = {
-      // required fields never have defaults
-      if (required && !forceOptional) {
-        TUPLE(REF("__") DOT "\\" APPLY LIT(rawName)) DOT "read" APPLYTYPE `type`
-      } else if (repeated && !forceOptional) {
-        TUPLE(REF("__") DOT "\\" APPLY LIT(rawName)) DOT "read" APPLYTYPE `type` DOT "orElse" APPLY(REF(PlayReads) DOT "pure" APPLY(`type` APPLY()))
-      } else {
-        if (defaultValue.isDefined && !forceOptional) {
-          TUPLE((REF("__") DOT "\\" APPLY LIT(rawName)) DOT "read" APPLYTYPE `type`) DOT "orElse" APPLY (REF(PlayReads) DOT "pure" APPLY defaultValue.get)
-        } else {
-          TUPLE((REF("__") DOT "\\" APPLY LIT(rawName)) DOT "readNullable" APPLYTYPE `type`)
-        }
-      }
-    }
-
-    val playValidator = {
-      def reads = constraints.validate(PlayPath DOT "read" APPLYTYPE `type`)
-      def validate =
-        REF("json") DOT "\\" APPLY(LIT(rawName)) DOT "validate" APPLYTYPE `type` APPLY(reads)
-      def validateOpt =
-        REF("json") DOT "\\" APPLY(LIT(rawName)) DOT "validateOpt" APPLYTYPE `type` APPLY(reads)
-      def validateOptWithDefault(defaultValue: Tree) =
-        REF("json") DOT "\\" APPLY(LIT(rawName)) DOT "validateOpt" APPLYTYPE `type` APPLY(reads) DOT "map" APPLY (REF("_") DOT "getOrElse" APPLY defaultValue)
-
-      if (required && !forceOptional) {
-        validate
-      } else if (repeated && !forceOptional) {
-        validateOptWithDefault(`type` APPLY())
-      } else {
-        if (defaultValue.isDefined && !forceOptional) {
-          validateOptWithDefault(defaultValue.get)
-        } else {
-          validateOpt
-        }
-      }
-    }
-  }
-
-  case class ObjectT(name: String, fields: Seq[FieldT], parentType: Option[String], comments: Seq[String], childTypes: Seq[ObjectT] = Nil, discriminator: Option[String] = None, discriminatorValue: Option[String] = None, serializeOnly: Boolean = false) extends GeneratedClass {
-    override def toString: String = parentType.fold(s"$name(${fields.mkString(", ")})")(parent => s"$name(${fields.mkString(" , ")}) extends $parent")
-
-    override def toTree(): Seq[Tree] = {
-      val actualFields = fields.filter(_.rawName != discriminator.getOrElse(""))
-      val params = actualFields.map(_.param)
-      val klass = if (childTypes.nonEmpty) {
-        if (params.nonEmpty) {
-          parentType.fold(TRAITDEF(name) withParents("RamlGenerated", "Product", "Serializable") := BLOCK(params))(parent =>
-            TRAITDEF(name) withParents(parent, "Product", "Serializable") := BLOCK(params)
-          )
-        } else {
-          parentType.fold((TRAITDEF(name) withParents("RamlGenerated", "Product", "Serializable")).tree)(parent =>
-            (TRAITDEF(name) withParents(parent, "Product", "Serializable")).tree
-          )
-        }
-      } else {
-        parentType.fold(CASECLASSDEF(name) withParents("RamlGenerated") withParams params)(parent =>
-          CASECLASSDEF(name) withParams params withParents parent
-        ).tree
-      }
-
-      val playFormat = if (discriminator.isDefined) {
-        Seq(
-          IMPORT("play.api.libs.json._"),
-
-          OBJECTDEF("playJsonFormat") withParents PLAY_JSON_FORMAT(name) withFlags Flags.IMPLICIT := BLOCK(
-            DEF("reads", PLAY_JSON_RESULT(name)) withParams PARAM("json", PlayJsValue) := BLOCK(
-              if (actualFields.size > 1) {
-                Seq(IMPORT("play.api.libs.functional.syntax._"),
-                  actualFields.map(_.playReader).reduce(_ DOT "and" APPLY _) DOT "apply" APPLY (REF(name) DOT "apply _") DOT "reads" APPLY REF("json"))
-              } else if (actualFields.size == 1) {
-                Seq(actualFields.head.playReader DOT "map" APPLY(REF(name) DOT "apply _") DOT "reads" APPLY REF("json"))
-              } else {
-                Seq(REF(name))
-              }
-            ),
-            DEF("writes", PlayJsObject) withParams PARAM("o", name) := {
-              REF(PlayJson) DOT "obj" APPLY
-                fields.map { field =>
-                  if (field.rawName == discriminator.get) {
-                    TUPLE(LIT(field.rawName), REF(PlayJson) DOT "toJsFieldJsValueWrapper" APPLY(PlayJson DOT "toJson" APPLY LIT(discriminatorValue.getOrElse(name))))
-                  } else {
-                    TUPLE(LIT(field.rawName), REF(PlayJson) DOT "toJsFieldJsValueWrapper" APPLY(PlayJson DOT "toJson" APPLY (REF("o") DOT field.rawName)))
-                  }
-                }
-            }
-          )
-        )
-      } else if (actualFields.nonEmpty && actualFields.exists(_.default.nonEmpty) && !actualFields.exists(f => f.repeated || f.omitEmpty || f.constraints.nonEmpty)) {
-        Seq(
-          IMPORT("play.api.libs.json._"),
-          IMPORT("play.api.libs.functional.syntax._"),
-          VAL("playJsonReader") withType PLAY_JSON_READS(name) := TUPLE(
-            actualFields.map(_.playReader).reduce(_ DOT "and" APPLY _)
-          ) APPLY (REF(name) DOT "apply _"),
-          VAL("playJsonWriter") withType PLAY_JSON_WRITES(name) := REF(PlayJson) DOT "writes" APPLYTYPE (name),
-          OBJECTDEF("playJsonFormat") withParents PLAY_JSON_FORMAT(name) withFlags Flags.IMPLICIT := BLOCK(
-            DEF("reads", PLAY_JSON_RESULT(name)) withParams PARAM("json", PlayJsValue) := BLOCK(
-              REF("playJsonReader") DOT "reads" APPLY(REF("json"))
-            ),
-            DEF("writes", PlayJsValue) withParams PARAM("o", name) := BLOCK(
-              REF("playJsonWriter") DOT "writes" APPLY REF("o")
-            )
-          )
-        )
-      } else if (actualFields.size > 22 || actualFields.exists(f => f.repeated || f.omitEmpty || f.constraints.nonEmpty) ||
-        actualFields.map(_.toString).exists(t => t.toString.startsWith(name) || t.toString.contains(s"[$name]"))) {
-        actualFields.map(_.constraints).requiredImports ++ Seq(
-          OBJECTDEF("playJsonFormat") withParents (if (serializeOnly) PLAY_JSON_WRITES(name) else PLAY_JSON_FORMAT(name)) withFlags Flags.IMPLICIT := BLOCK(
-            if (serializeOnly) {
-              Seq()
-            } else  Seq(DEF("reads", PLAY_JSON_RESULT(name)) withParams PARAM("json", PlayJsValue) := {
-                BLOCK(
-                  actualFields.map { field =>
-                    VAL(field.name) := field.playValidator
-                  } ++ Seq(
-                    VAL("_errors") := SEQ(actualFields.map(f => TUPLE(LIT(f.rawName), REF(f.name)))) DOT "collect" APPLY BLOCK(
-                      CASE(REF(s"(field, e:$PlayJsError)")) ==> (REF("e") DOT "repath" APPLY (REF(PlayPath) DOT "\\" APPLY REF("field"))) DOT s"asInstanceOf[$PlayJsError]"),
-                    IF(REF("_errors") DOT "nonEmpty") THEN (
-                      REF("_errors") DOT "reduceOption" APPLYTYPE PlayJsError APPLY (REF("_") DOT "++" APPLY REF("_")) DOT "getOrElse" APPLY (REF("_errors") DOT "head")
-                      ) ELSE (
-                      REF(PlayJsSuccess) APPLY (REF(name) APPLY
-                        actualFields.map { field =>
-                          REF(field.name) := (REF(field.name) DOT "get")
-                        }))
-                  )
-                )
-            }) ++ Seq(
-              DEF("writes", PlayJsValue) withParams PARAM("o", name) := BLOCK(
-                actualFields.withFilter(_.name != AdditionalProperties).map { field =>
-                  val serialized = REF(PlayJson) DOT "toJson" APPLY (REF("o") DOT field.name)
-                  if (field.omitEmpty && field.repeated && !field.forceOptional) {
-                    VAL(field.name) := IF(REF("o") DOT field.name DOT "nonEmpty") THEN (
-                      serialized
-                      ) ELSE (
-                      PlayJsNull
-                      )
-                  } else if(field.omitEmpty && !field.repeated && !builtInTypes.contains(field.`type`.toString())) {
-                    // earlier "require" check ensures that we won't see a field w/ omitEmpty that is not optional.
-                    // see buildTypes
-                    VAL(field.name) := serialized MATCH(
-                      // avoid serializing JS objects w/o any fields
-                      CASE(ID("obj") withType (PlayJsObject),
-                        IF(REF("obj.fields") DOT "isEmpty")) ==> PlayJsNull,
-                      CASE(ID("rs")) ==> REF("rs")
-                    )
-                  } else {
-                    VAL(field.name) := serialized
-                  }
-                } ++
-                  Seq(
-                    REF(PlayJsObject) APPLY (SEQ(
-                      actualFields.withFilter(_.name != AdditionalProperties).map { field =>
-                        TUPLE(LIT(field.rawName), REF(field.name))
-                      }) DOT "filter" APPLY (REF("_._2") INFIX("!=") APPLY PlayJsNull) DOT("++") APPLY(
-                        actualFields.find(_.name == AdditionalProperties).fold(REF("Seq") DOT "empty") { extraPropertiesField =>
-                        REF("o.additionalProperties") DOT "fields"
-                      })
-                    )
-                  )
-              )
-            )
-          )
-        )
-      } else {
-        Seq(VAL("playJsonFormat") withFlags Flags.IMPLICIT := REF("play.api.libs.json.Json") DOT "format" APPLYTYPE (name))
-      }
-
-      val defaultFields = fields.withFilter(_.paramTypeValue.nonEmpty).flatMap { f =>
-        val (dType, dValue) = f.paramTypeValue.get
-        val fieldName = (
-          if (f.name.contains("-")) underscoreToCamel(f.name.replace('-', '_')) else f.name
-        ).replace("`", "").capitalize
-        Seq(VAL(s"Default${fieldName}") withType(dType) := dValue)
-      }
-
-      val defaultInstance: Seq[Tree] =
-        if (fields.forall(f => f.defaultValue.nonEmpty || f.forceOptional || (f.repeated && !f.required))) {
-          Seq(VAL("Default") withType (name) := REF(name) APPLY())
-        } else Nil
-
-      val obj = if (childTypes.isEmpty || serializeOnly) {
-        (OBJECTDEF(name)) := BLOCK(
-          playFormat ++ defaultFields ++ defaultInstance ++ fields.flatMap { f =>
-            f.constraints.flatMap(_.withFieldLimit(f).limitField)
-          }
-        )
-      } else if (discriminator.isDefined) {
-        val childDiscriminators: Map[String, ObjectT] = childTypes.map(ct => ct.discriminatorValue.getOrElse(ct.name) -> ct)(collection.breakOut)
-        OBJECTDEF(name) := BLOCK(
-          Seq(OBJECTDEF("PlayJsonFormat") withParents PLAY_JSON_FORMAT(name) withFlags Flags.IMPLICIT := BLOCK(
-            DEF("reads", PLAY_JSON_RESULT(name)) withParams PARAM("json", PlayJsValue) := {
-              TUPLE(REF("json") DOT "\\" APPLY LIT(discriminator.get)) DOT "validate" APPLYTYPE (StringClass) MATCH (
-                childDiscriminators.map { case (k, v) =>
-                  CASE(PlayJsSuccess APPLY(LIT(k), REF("_"))) ==> (REF("json") DOT "validate" APPLYTYPE (v.name))
-                } ++
-                  Seq(
-                    CASE(WILDCARD) ==> (REF(PlayJsError) APPLY (REF(PlayValidationError) APPLY(LIT("error.expected.jsstring"), LIT(s"expected one of (${childDiscriminators.keys.mkString(", ")})"))))
-                  )
-                )
-            },
-            DEF("writes", PlayJsValue) withParams PARAM("o", name) := BLOCK(
-              REF("o") MATCH
-                childDiscriminators.map { case (k, v) =>
-                  CASE(REF(s"f:${v.name}")) ==> (REF(PlayJson) DOT "toJson" APPLY REF("f") APPLY(REF(v.name) DOT "playJsonFormat"))
-                }
-            )
-          )) ++ defaultFields ++ defaultInstance
-        )
-      } else {
-        System.err.println(s"[WARNING] $name uses subtyping but has no discriminator. If it is not a union type when it is" +
-          " used, it will not be able to be deserialized at this time")
-        OBJECTDEF(name) := BLOCK(defaultFields ++ defaultInstance)
-      }
-
-      val commentBlock = comments ++ actualFields.map(_.comment)(collection.breakOut)
-      Seq(klass.withDoc(commentBlock)) ++ childTypes.flatMap(_.toTree()) ++ Seq(obj)
-    }
-  }
-
-  case class StringT(name: String, defaultValue: Option[String]) extends GeneratedClass {
-    override def toTree(): Seq[treehugger.forest.Tree] = Seq.empty[Tree]
-  }
-
-  case class UnionT(name: String, childTypes: Seq[GeneratedClass], comments: Seq[String]) extends GeneratedClass {
-    override def toString: String = s"Union($name, $childTypes)"
-
-    override def toTree(): Seq[Tree] = {
-      val base = (TRAITDEF(name) withParents("RamlGenerated", "Product", "Serializable")).tree.withDoc(comments)
-      val childJson: Seq[GenericApply] = childTypes.map { child =>
-        REF("json") DOT s"validate" APPLYTYPE (child.name)
-      }
-
-      val obj = OBJECTDEF(name) := BLOCK(
-        OBJECTDEF("playJsonFormat") withParents PLAY_JSON_FORMAT(name) withFlags Flags.IMPLICIT := BLOCK(
-          DEF("reads", PLAY_JSON_RESULT(name)) withParams PARAM("json", PlayJsValue) := BLOCK(
-            childJson.reduce((acc, next) => acc DOT "orElse" APPLY next)
-          ),
-          DEF("writes", PlayJsValue) withParams PARAM("o", name) := BLOCK(
-            REF("o") MATCH
-              childTypes.map { child =>
-                CASE(REF(s"f:${child.name}")) ==> (REF(PlayJson) DOT "toJson" APPLY REF("f") APPLY(REF(child.name) DOT "playJsonFormat"))
-              }
-          )
-        )
-      )
-      val children = childTypes.flatMap {
-        case s: StringT =>
-          Seq[Tree](
-            CASECLASSDEF(s.name) withParents name withParams s.defaultValue.fold(PARAM("value", StringClass).tree){ defaultValue =>
-              PARAM("value", StringClass) := LIT(defaultValue)
-            },
-            OBJECTDEF(s.name) := BLOCK(
-              Seq(OBJECTDEF("playJsonFormat") withParents PLAY_JSON_FORMAT(s.name) withFlags Flags.IMPLICIT := BLOCK(
-                DEF("reads", PLAY_JSON_RESULT(s.name)) withParams PARAM("json", PlayJsValue) := BLOCK(
-                  REF("json") DOT "validate" APPLYTYPE StringClass DOT "map" APPLY (REF(s.name) DOT "apply")
-                ),
-                DEF("writes", PlayJsValue) withParams PARAM("o", s.name) := BLOCK(
-                  REF(PlayJsString) APPLY (REF("o") DOT "value")
-                )
-              )) ++ s.defaultValue.map{ defaultValue =>
-                VAL("DefaultValue") withType(s.name) := REF(s.name) APPLY()
-              }
-            )
-          )
-        case t => t.toTree()
-      }
-      Seq(base) ++ children ++ Seq(obj)
-    }
   }
 
   @tailrec def libraryTypes(libraries: List[Library], result: Set[TypeDeclaration] = Set.empty): Set[TypeDeclaration] = {
@@ -744,7 +192,7 @@ object RamlTypeGenerator {
           s"field $fieldOwner.${field.name()} specifies omitEmpty but is required or provides a default value")
 
         def arrayType(a: ArrayTypeDeclaration): Type =
-          if (scala.util.Try[Boolean](a.uniqueItems()).getOrElse(false)) SetClass else SeqClass
+          if (scala.util.Try[Boolean](a.uniqueItems()).getOrElse(false)) SetClass else backend.SeqClass
 
         field match {
           case a: ArrayTypeDeclaration =>
@@ -908,10 +356,14 @@ object RamlTypeGenerator {
   }
 
   def apply(models: Seq[RamlModelResult], pkg: String): Map[String, Tree] = {
+    // Front end: Parsed type declarations.
     val typeDeclarations = allTypes(models)
+
+    // Middle end: Intermediate representation.
     val typeTable = buildTypeTable(typeDeclarations)
     val types = buildTypes(typeTable, typeDeclarations)
 
+    // Back end: Code generation with Treehugger.
     generateBuiltInTypes(pkg) ++ types.map { tpe =>
       val tree = tpe.toTree()
       if (tree.nonEmpty) {
