+++ type-generator/src/main/scala/mesosphere/raml/ir/StringT.scala
@@ -1,0 +1,8 @@
+package mesosphere.raml.ir
+
+import treehugger.forest.Tree
+
+case class StringT(name: String, defaultValue: Option[String]) extends GeneratedClass {
+  // TODO(karsten): This is actually part of the back end.
+  override def toTree(): Seq[treehugger.forest.Tree] = Seq.empty[Tree]
+}
